tests/wpt/tests/css/css-conditional/container-queries/scroll-state/scroll-state-snapped-none.html

<!DOCTYPE html>
<title>@container: scroll-state(snapped: none)</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#snapped">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<style>
  #no-snap { container-type: scroll-state; }
  @container scroll-state(snapped: none) {
    #target { --snapped-none: true }
  }
  @container not scroll-state(snapped) {
    #target { --snapped-boolean: true }
  }
</style>
<div id="no-snap">
  <div id="target"></div>
</div>
<script>
  setup(() => assert_implements_container_queries());

  test(() => {
    assert_equals(getComputedStyle(target).getPropertyValue("--snapped-none"), "true");
    assert_equals(getComputedStyle(target).getPropertyValue("--snapped-boolean"), "true");
  }, "Check scroll-state(snapped:none) and scroll-state(snapped) for non-snapping container");

</script>
